Comments at https://github.com/PipedreamHQ/pipedream/pull/1759\#pullrequestreview-799396028 were addressed
